Change type of config. to uint64
