+++ amo/config.go
@@ -1,0 +1,21 @@
+package amo
+
+type AMOAppConfig struct {
+	MaxValidators   uint64 `json:"max_validators"`
+	WeightValidator uint64 `json:"weight_validator"`
+	WeightDelegator uint64 `json:"weight_delegator"`
+
+	MinStakingUnit string `json:"min_staking_unit"`
+
+	BlkReward uint64 `json:"blk_reward"`
+	TxReward  uint64 `json:"tx_reward"`
+
+	PenaltyRatioM float64 `json:"penalty_ratio_m"` // malicious validator
+	PenaltyRatioL float64 `json:"penalty_ratio_l"` // lazy validators
+
+	LazinessCounterWindow int64   `json:"laziness_counter_window"`
+	LazinessThreshold     float64 `json:"laziness_threshold"`
+
+	BlockBoundTxGracePeriod uint64 `json:"block_bound_tx_grace_period"`
+	LockupPeriod            uint64 `json:"lockup_period"`
+}
